tools/sdi12-logger/requirements.txt
pyserial>=3.5
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
